# docs/swagger.yaml
definitions:
...
  model.Cart:
    properties:
      ProductId:
        type: integer
      id:
        type: integer
      qty:
        type: integer
      userid:
        type: integer
    type: object
  model.CartResponse:
    properties:
      data:
        $ref: '#/definitions/model.Cart'
      message:
        type: string
      status:
        type: integer
    type: object
  model.CartsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Cart'
        type: array
      message:
        type: string
      status:
        type: integer
    type: object
  model.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
  model.Product:
    properties:
      category:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      storeId:
        type: integer
    type: object
  model.ProductResponse:
    properties:
      data:
        $ref: '#/definitions/model.Product'
      message:
        type: string
      status:
        type: integer
    type: object
...
  model.Store:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      userId:
        type: integer
    type: object
  model.StoreResponse:
    properties:
      data:
        $ref: '#/definitions/model.Store'
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /cart:
    get:
      description: display all cart items of users who are currently logged in.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CartsResponse'
      summary: Get all cart items.
      tags:
      - Cart
    post:
      description: insert product to cart belongs to the user who is currently logged
        in.
      parameters:
      - description: cart's data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.Cart'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CartResponse'
      summary: insert cart.
      tags:
      - Cart
  /cart/{cartId}:
    delete:
      description: delete display all cart items of users who are currently logged
        in.
      parameters:
      - description: cartId
        in: path
        name: cartId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: delete cart item.
      tags:
      - Cart
    post:
      description: update cart belongs to the user who is currently logged in.
      parameters:
      - description: cart's data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.Cart'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: update cart.
      tags:
      - Cart
...
  /product:
    post:
      description: insert products sold by logged in users.
      parameters:
      - description: product's data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProductResponse'
      summary: insert product.
      tags:
      - Products
  /product/{productid}:
    delete:
      description: delete products sold by logged in users.
      parameters:
      - description: productid
        in: path
        name: productid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: delete prodduct.
      tags:
      - Products
    put:
      description: product updates sold by logged in users.
      parameters:
      - description: productid
        in: path
        name: productid
        required: true
        type: string
      - description: product's data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProductResponse'
      summary: update product's data.
      tags:
      - Products
...
  /store:
    delete:
      description: delete the store belonging to the user who is currently logged
        in.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: delete store.
      tags:
      - Stores
    post:
      description: added the store of the currently logged in user.
      parameters:
      - description: store's data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.Store'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StoreResponse'
      summary: insert store.
      tags:
      - Stores
    put:
      description: update the store of the currently logged in user.
      parameters:
      - description: store's data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.Store'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StoreResponse'
      summary: update store.
      tags:
      - Stores
...
swagger: "2.0"